+++ SoundBite/SBSystemInfoViewController.m
@@ -99,12 +99,12 @@
         
 	NSUInteger row = [indexPath row];
 
-	NSString *xpath = [NSString stringWithFormat:@"/Envelope/Body/showSystemInfoResponse/return/data[%lu]/name", row+1];
+	NSString *xpath = [NSString stringWithFormat:@"/Envelope/Body/showSystemInfoResponse/return/data[%lu]/name", (unsigned long)row+1];
 	NSArray *nodes = [self.sbSoap.doc nodesForXPath:xpath error:nil];
     cell.textLabel.text = [nodes[0] stringValue];
 	//cell.nameLabel.text = [nodes[0] stringValue];
 	
-	xpath = [NSString stringWithFormat:@"/Envelope/Body/showSystemInfoResponse/return/data[%lu]/value", row+1];
+	xpath = [NSString stringWithFormat:@"/Envelope/Body/showSystemInfoResponse/return/data[%lu]/value", (unsigned long)row+1];
     nodes = [self.sbSoap.doc nodesForXPath:xpath error:nil];
     cell.detailTextLabel.text = [nodes[0] stringValue];
